move dacsspace to where it was
