+++ dacsspace/__init__.py
@@ -1,0 +1,6 @@
+__title__ = 'DACSspace'
+__version__ = '0.1.0'
+__license__ = 'MIT'
+__copyright__ = 'Copyright 2022 Rockefeller Archive Center'
+
+from .dacsspace import DACSspace  # noqa: F401
